# Fix fast_reader parsing of RDB headers with names specified and performance issues