--- a/CHANGES.rst
+++ b/CHANGES.rst
@@ -171,11 +171,6 @@ astropy.extern
 astropy.io.ascii
 ^^^^^^^^^^^^^^^^
 
-- Fix failure of ASCII ``fast_reader`` to handle ``names``, ``include_names``,
-  ``exclude_names`` arguments for ``RDB`` formatted tables; homogenised checks
-  and exceptions for invalif ``names`` arguments; improved performance when
-  parsing tables with duplicate column names. [#10306]
-
 astropy.io.fits
 ^^^^^^^^^^^^^^^
 
@@ -295,6 +290,11 @@ astropy.io.ascii
 - Functional Units can now be processed in CDS-tables. [#9971]
 
 - Allow reading in ASCII tables which have duplicate column names. [#9939]
+
+- Fixed failure of ASCII ``fast_reader`` to handle ``names``, ``include_names``,
+  ``exclude_names`` arguments for ``RDB`` formatted tables. Homogenised checks
+  and exceptions for invalid ``names`` arguments. Improved performance when
+  parsing "wide" tables with many columns. [#10306]
 
 - Added type validation of key arguments in calls to ``io.ascii.read()`` and
   ``io.ascii.write()`` functions. [#10005]