--- a/astropy/io/ascii/cparser.pyx
+++ b/astropy/io/ascii/cparser.pyx
@@ -308,7 +308,7 @@ cdef class CParser:
         self.tokenizer.source = self.source_bytes
         self.tokenizer.source_len = <size_t>len(self.source_bytes)
 
-    def read_header(self, deduplicate=True):
+    def read_header(self, deduplicate=True, filter_names=True):
         self.tokenizer.source_pos = 0
 
         # header_start is a valid line number
@@ -321,9 +321,9 @@ cdef class CParser:
             self.header_names = []
             name = ''
 
-            for i in range(self.tokenizer.output_len[0]): # header is in first col string
-                c = self.tokenizer.output_cols[0][i] # next char in header string
-                if not c: # zero byte -- field terminator
+            for i in range(self.tokenizer.output_len[0]):  # header is in first col string
+                c = self.tokenizer.output_cols[0][i]       # next char in header string
+                if not c:  # zero byte -- field terminator
                     if name:
                         # replace empty placeholder with ''
                         self.header_names.append(name.replace('\x01', ''))
@@ -341,17 +341,17 @@ cdef class CParser:
             if tokenize(self.tokenizer, -1, 1, 0) != 0:
                 self.raise_error("an error occurred while tokenizing the first line of data")
             self.width = 0
-            for i in range(self.tokenizer.output_len[0]): # header is in first col string
+            for i in range(self.tokenizer.output_len[0]):  # header is in first col string
                 # zero byte -- field terminator
                 if not self.tokenizer.output_cols[0][i]:
                     # ends valid field
                     if i > 0 and self.tokenizer.output_cols[0][i - 1]:
                         self.width += 1
-                    else: # end of line
+                    else:  # end of line
                         break
-            if self.width == 0: # no data
+            if self.width == 0:  # no data
                 raise core.InconsistentTableError('No data lines found, C reader '
-                                            'cannot autogenerate column names')
+                                                  'cannot autogenerate column names')
             # auto-generate names
             self.header_names = ['col{0}'.format(i + 1) for i in range(self.width)]
 
@@ -362,9 +362,9 @@ cdef class CParser:
 
         # self.use_cols should only contain columns included in output
         self.use_cols = set(self.names)
-        if self.include_names is not None:
+        if filter_names and self.include_names is not None:
             self.use_cols.intersection_update(self.include_names)
-        if self.exclude_names is not None:
+        if filter_names and self.exclude_names is not None:
             self.use_cols.difference_update(self.exclude_names)
 
         self.width = <int>len(self.names)