--- a/astropy/io/ascii/tests/test_read.py
+++ b/astropy/io/ascii/tests/test_read.py
@@ -1535,40 +1535,45 @@ def test_deduplicate_names_basic(rdb, fast_reader):
     assert dat.colnames == ['a', 'a_2', 'a_1', 'a_3', 'a_4']
     assert len(dat) == 2
 
-    if rdb is False or fast_reader is False:
-        dat = ascii.read(lines, fast_reader=fast_reader, include_names=['a', 'a_2', 'a_3'])
-        assert len(dat) == 2
-        assert dat.colnames == ['a', 'a_2', 'a_3']
-        assert np.all(dat['a'] == [1, 10])
-        assert np.all(dat['a_2'] == [2, 20])
-        assert np.all(dat['a_3'] == [4, 40])
+    dat = ascii.read(lines, fast_reader=fast_reader, include_names=['a', 'a_2', 'a_3'])
+    assert len(dat) == 2
+    assert dat.colnames == ['a', 'a_2', 'a_3']
+    assert np.all(dat['a'] == [1, 10])
+    assert np.all(dat['a_2'] == [2, 20])
+    assert np.all(dat['a_3'] == [4, 40])
 
-        dat = ascii.read(lines, fast_reader=fast_reader,
-                         names=['b1', 'b2', 'b3', 'b4', 'b5'],
-                         include_names=['b1', 'b2', 'b4'])
-        assert len(dat) == 2
-        assert dat.colnames == ['b1', 'b2', 'b4']
-        assert np.all(dat['b1'] == [1, 10])
-        assert np.all(dat['b2'] == [2, 20])
-        assert np.all(dat['b4'] == [4, 40])
+    dat = ascii.read(lines, fast_reader=fast_reader,
+                     names=['b1', 'b2', 'b3', 'b4', 'b5'],
+                     include_names=['b1', 'b2', 'a_4', 'b4'])
+    assert len(dat) == 2
+    assert dat.colnames == ['b1', 'b2', 'b4']
+    assert np.all(dat['b1'] == [1, 10])
+    assert np.all(dat['b2'] == [2, 20])
+    assert np.all(dat['b4'] == [4, 40])
+
+    dat = ascii.read(lines, fast_reader=fast_reader,
+                     names=['b1', 'b2', 'b3', 'b4', 'b5'],
+                     exclude_names=['b3', 'b5', 'a_3', 'a_4'])
+    assert len(dat) == 2
+    assert dat.colnames == ['b1', 'b2', 'b4']
+    assert np.all(dat['b1'] == [1, 10])
+    assert np.all(dat['b2'] == [2, 20])
+    assert np.all(dat['b4'] == [4, 40])
 
 
-@pytest.mark.xfail
-def test_include_names_rdb_fast_fail0():
-    """Test that selecting column names with `include_names` fails for the rdb format for fast reader.
-    This is not desired but reflects a current known limitation.
+def test_include_names_rdb_fast():
+    """Test that selecting column names with `include_names` works for the rdb format
+    with fast reader. This is testing the fix for a bug identified in #9939.
     """
     lines = _get_lines(True)
     lines[0] = 'a\ta_2\ta_1\ta_3\ta_4'
-
     dat = ascii.read(lines, fast_reader='force', include_names=['a', 'a_2', 'a_3'])
     assert len(dat) == 2
 
 
-@pytest.mark.xfail
-def test_include_names_rdb_fast_fail1():
-    """Test that `include_names` fails with duplicate column names for the rdb format for fast reader.
-    This is not desired but reflects a current known limitation.
+def test_deduplicate_names_rdb_fast1():
+    """Test that selecting column names with `include_names` works for the rdb format
+    with fast reader and duplicate column names. Testing fix for a bug identified in #9939.
     """
     lines = _get_lines(True)
 
@@ -1576,10 +1581,9 @@ def test_include_names_rdb_fast_fail1():
     assert len(dat) == 2
 
 
-@pytest.mark.xfail
-def test_include_names_rdb_fast_fail2():
-    """Test that `include_names` fails with duplicate column names for the rdb format for fast reader.
-    This is not desired but reflects a current known limitation.
+def test_deduplicate_names_rdb_fast2():
+    """Test that selecting column names with `include_names` works for the rdb format
+    with fast reader and duplicate column names. Testing fix for a bug identified in #9939.
     """
     lines = _get_lines(True)