### Description
#### for 4.2
Proposed fix for a `fast_reader` bug with `format='rdb'` identified in #9939 that occurs however unrelated to duplicate column names, whenever any of the `names`, `include_names` or `exclude_names` arguments are used. E.g.
```
>>> lines = ['a_0\ta_1\ta_2', 'N\tN\tN', '1\t2\t3', '10\t20\t30']
>>> ascii.read(lines, fast_reader='force', include_names=['a_1', 'a_2'])
<Table length=3>
a_1  a_2 
str2 str2
---- ----
   N    N
   2    3
  20   30
>>> ascii.read(lines, fast_reader='force', include_names=['a_1', 'a_2'], format='rdb')
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/sw/lib/python3.8/site-packages/astropy/io/ascii/ui.py", line 304, in read
    dat = fast_reader_rdr.read(table)
  File "/sw/lib/python3.8/site-packages/astropy/io/ascii/fastbasic.py", line 118, in read
    conversion_info = self._read_header()
  File "/sw/lib/python3.8/site-packages/astropy/io/ascii/fastbasic.py", line 341, in _read_header
    raise core.InconsistentTableError('RDB header mismatch between number of '
astropy.io.ascii.core.InconsistentTableError: RDB header mismatch between number of column names and column types
```

#### `names` fixes
The reason is `rdb` is using a two-line header format with line 2 storing the type information, both of which are parsed using `CParser.read_header` that has applied to the 2nd line the full set of renaming and selection operations intended for the 1st line with the actual column names.

The fix follows the discussion in https://github.com/astropy/astropy/pull/9939#issuecomment-622574610 by always setting the `types` entries from `header_names`, saving custom `names` once set, and disabling the `in|exclude_names` filtering on the 2nd header line.

#### Timings
I have made some performance tests with a 4096-column table (otherwise equivalent to the test `lines` from #9939 with two rows of integer data):

with unique column names - 4.0.1:
`ascii.read(lines, fast_reader='force')`: 230 ms
`ascii.read(lines, fast_reader=False)`: 260 ms
`ascii.read(lines, fast_reader=False, names=alt_names)`: 35 s

master with #9939 merged:
`ascii.read(lines, fast_reader='force')`: 240 ms
`ascii.read(lines, fast_reader=False)`: 7 s
`ascii.read(lines, fast_reader=False, names=alt_names)`: 17 s
with duplicate names:
`ascii.read(lines, fast_reader=False)`: 9 s

`fast_reader` fixes from this PR:
`ascii.read(lines, fast_reader='force')`: 245 ms
`ascii.read(lines, fast_reader=False)`: 7 s
`ascii.read(lines, fast_reader='force', names=alt_names)`: 316 ms
with duplicate names:
`ascii.read(lines, fast_reader='force')`: 2.7 s
`ascii.read(lines, fast_reader='force', names=alt_names)`: 320 ms

Timings for `include_names` range from ca. 40 ms for up to 2 % of columns selected, to about the same times as with `names` for most of the columns selected.

The primary motivation for these timings were to check if the additional header reading in the fix impacted performance (which it does at the ~1 % level), but it also revealed some changes with  the Python parser:

1. The column renaming (for `names=alt_names`) has significantly gained in performance thanks to the new `rename_columns` implementation in #9939.

2. The standard parsing has nonetheless taken a 20x performance hit over the `fast_reader`.

#### Additional performance tuning
edec8f7 was an attempt to address some of the performance penalties in constructing unique names for `CParser`:

1. by using its own instance of `_deduplicate_names` pulled into `cparser.pyx`; this does not add any string operations on the C level, but since it's giving about 30 % speedup basically for free, it's perhaps worth the slight increase in code complexity.

2. skipping the check for duplicate names entirely if `names` have been passed as input argument, since in that case `header_names` will never be used anyway.

#### Checks and tests for invalid `names` arguments
In the 2nd case above it is of course the user's responsibility to provide unique column names, but invalid `names` have raised different exceptions in various different parts of the code. 4334b08 adds checks directly on calling `ascii.read` both for the Python and fast readers and ensures consistent exceptions with those raised in `Table` plus tests. With the default settings with `guess` enabled the exceptions still tend to be buried among the output on all the different formats tried, but they should at least be a bit easier to identify.

#### Performance fixes for Python parser
The reason for the dramatic slowdown on regular input (with valid, unique column names) turned out to be the check for `colnames_uniq != table.colnames`; since the former is a list and the latter a tuple at that point, the comparison evaluated always `False` enforcing a call to `rename_columns`.
In addition, with `names` specified, `table.colnames` would first be renamed to `colnames_uniq` and then to `names` from input. Changed analogously to the handling in `CParser`. This improves the timings to
unique column names:
`ascii.read(lines, fast_reader=False)`: 280 ms
`ascii.read(lines, fast_reader=False, names=alt_names)`: 5.8 s
duplicate names:
`ascii.read(lines, fast_reader=False)`: 9 s
`ascii.read(lines, fast_reader=False, names=alt_names)`: 6.5 s