@pllim 2020-05-05
Thanks! You can hold off on adding a change log until after 4.1 is released.

Does this need new benchmarks at https://github.com/astropy/astropy-benchmarks ?