@dhomeier 2020-05-05
> Does this need new benchmarks at https://github.com/astropy/astropy-benchmarks ?

I had only noticed that repo in reviewing `CONTRIBUTING.md` right before this PR, but never added anything there. I think there are some benchmarks that could be added, will look into that once I have checked how those benchmarks are organised and executed.