@dhomeier 2020-05-05
Well I had been wondering if any of the `io.ascii` benchmarks from https://github.com/astropy/astropy-notebooks had made it there, so I am certainly going to have a closer look.