@hamogu 2020-05-05
While I'm the deputy sub-package maintainer, I can't say that I have more than a vague memory of hearing about astropy benchmarks a few years ago. I certainly did not know where they live, how they work, or what is being tested there, so please don't expect input from me @dhomeier ;-)