@taldcroft 2020-05-21
@dhomeier - sorry.. the title of this made me think the scope was basically the fast RDB parsing, which as you know I consider to be a low priority so I didn't have a look until now. (In fact if you had not spent so much work on this I might consider just dropping RDB support entirely from the fast reader - we are thinking about *reducing* some functionality for 4.2 to simplify code and reduce "technical debt").

On the other hand, the performance regression and bug you found in #9939 is critical and urgent.  Great catch! A separate PR or issue like "Serious performance regression from duplicate column handling" would have gotten a response 16 days ago. My bad, but I have many things going on and have to prioritize.

So... @bsipocz @pllim - what's happening with the 4.1 release? Can we get a bug fix / performance regression fix in for 4.1?  This can be literally changing `table.colnames` to `list(table.colnames)` in one line. This impacts performance for every call to the Python reader. This could go in a 4.1.1 release of course.