@taldcroft 2020-05-21
About the benchmarking, see: https://www.astropy.org/astropy-benchmarks/#/

There are no obvious regressions there, so clearly we need to add tests and then pay attention (or have some alerts sent out for new regressions).