@taldcroft 2020-05-21
A simple thing now is just pulling out 4362f199f into a separate performance-regression-fix PR.