@taldcroft 2020-05-21
@astrofrog - OK.  I'll actually review this whole PR which is a bugfix + performance fix. If there are issues that can't be resolved quickly then I'll just do that one commit.