@astrofrog 2020-05-22
@taldcroft - I'm going to try and push ahead with the RC today, how soon can you review this? :grin: