@astrofrog 2020-05-22
Note that the changelog entry will need to be moved to the v4.1 section