@taldcroft 2020-05-22
See: https://github.com/astropy/astropy-benchmarks/issues/93