@astrofrog 2020-05-22
I think the 32-bit failure can be ignored and has been fixed in master in the mean time