@dhomeier 2020-05-22
Thanks all for getting this speedily through the final review. Perhaps I should have marked the performance impact more prominently, but I thought there was consensus not to let this interfere with the feature freeze for 4.1, which seemed stressful enough as it was. In any case it did not return wrong results, and only hit when `fast_reader=False` was explicitly selected or needed to handle duplicate columns or `*names`.
But 4362f19 would have only helped with the specific case where custom `names` have been provided, so I am glad it made it in as a whole fix!